docker_hailo8_service/examples/client_app/requirements.txt
# 客户端应用依赖
aiohttp==3.8.5
requests==2.31.0
Pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0.76
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.3.0
loguru==0.7.0
python-dotenv==1.0.0
asyncio-mqtt==0.13.0